Relocate 3 presenters
GetCompanyTransactionsPresenter, GetStatisticsPresenter, ShowMyCooperationsPresenter
