+++ arbeitszeit_flask/company/routes.py
@@ -87,8 +87,6 @@ from arbeitszeit_web.create_draft import (
     CreateDraftPresenter,
     GetPrefilledDraftDataPresenter,
 )
-from arbeitszeit_web.get_company_transactions import GetCompanyTransactionsPresenter
-from arbeitszeit_web.get_statistics import GetStatisticsPresenter
 from arbeitszeit_web.presenters.company_purchases_presenter import (
     CompanyPurchasesPresenter,
 )
@@ -99,12 +97,16 @@ from arbeitszeit_web.presenters.file_plan_with_accounting_presenter import (
 from arbeitszeit_web.presenters.get_company_summary_presenter import (
     GetCompanySummarySuccessPresenter,
 )
+from arbeitszeit_web.presenters.get_company_transactions_presenter import (
+    GetCompanyTransactionsPresenter,
+)
 from arbeitszeit_web.presenters.get_coop_summary_presenter import (
     GetCoopSummarySuccessPresenter,
 )
 from arbeitszeit_web.presenters.get_plan_summary_company_presenter import (
     GetPlanSummaryCompanyPresenter,
 )
+from arbeitszeit_web.presenters.get_statistics_presenter import GetStatisticsPresenter
 from arbeitszeit_web.presenters.hide_plan_presenter import HidePlanPresenter
 from arbeitszeit_web.presenters.list_all_cooperations_presenter import (
     ListAllCooperationsPresenter,
@@ -113,6 +115,9 @@ from arbeitszeit_web.presenters.list_plans_presenter import ListPlansPresenter
 from arbeitszeit_web.presenters.query_companies_presenter import QueryCompaniesPresenter
 from arbeitszeit_web.presenters.show_a_account_details_presenter import (
     ShowAAccountDetailsPresenter,
+)
+from arbeitszeit_web.presenters.show_my_cooperations_presenter import (
+    ShowMyCooperationsPresenter,
 )
 from arbeitszeit_web.presenters.show_my_plans_presenter import ShowMyPlansPresenter
 from arbeitszeit_web.presenters.show_p_account_details_presenter import (
@@ -129,7 +134,6 @@ from arbeitszeit_web.request_cooperation import (
     RequestCooperationController,
     RequestCooperationPresenter,
 )
-from arbeitszeit_web.show_my_cooperations import ShowMyCooperationsPresenter
 from arbeitszeit_web.url_index import UrlIndex
 
 from .blueprint import CompanyRoute
