+++ tests/presenters/test_get_company_transactions.py
@@ -11,7 +11,9 @@ from arbeitszeit.use_cases.get_company_transactions import (
     GetCompanyTransactionsResponse,
     TransactionInfo,
 )
-from arbeitszeit_web.get_company_transactions import GetCompanyTransactionsPresenter
+from arbeitszeit_web.presenters.get_company_transactions_presenter import (
+    GetCompanyTransactionsPresenter,
+)
 from tests.datetime_service import FakeDatetimeService
 from tests.translator import FakeTranslator
