+++ tests/presenters/test_get_statistics_presenter.py
@@ -5,7 +5,7 @@ from unittest import TestCase
 
 from arbeitszeit.entities import PayoutFactor
 from arbeitszeit.use_cases.get_statistics import StatisticsResponse
-from arbeitszeit_web.get_statistics import GetStatisticsPresenter
+from arbeitszeit_web.presenters.get_statistics_presenter import GetStatisticsPresenter
 from tests.datetime_service import FakeDatetimeService
 
 from ..translator import FakeTranslator
